requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
pyyaml==6.0.1
langchain==0.1.0
langchain-google-genai==0.0.8
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
psutil==5.9.6